refactor: clarify credential preferences workflow path
